ether: defconfig: Enable WLAN_RX_WAKELOCK
Not having this causes serious delays in wifi transfers when the
phone sleeps.  These delays appear to cause power drain, perhaps
because the wifi is left in a higher power state but no data is
being pulled from the queue to let it drop back to a lower power
state?

Change-Id: I0097b6ca5d00ead56e84fbdb81064730d52b25c9